Add support for error handling and return PrideUser object in createAccount()
This commit adds support for error handling to the createAccount() method in the Query class. The method now returns a PrideUser object, which contains the generated ID and Minecraft UUID of the newly created account. This will allow for more efficient and accurate handling of user information in the future.

In addition to the return value, the createAccount() method now includes a Map of parameters, which includes the generated ID and Minecraft UUID of the account. These parameters are used in the Cypher query that creates the account and its relationships in the Neo4j database. The query has also been updated to return the relationships between the nodes, so that information about the account can be more easily accessed and utilized in the future.

Finally, the BuildJson class has been updated to include an error() method, which creates a JSON string with a status of "error" and a message detailing the error. This will allow for more precise and efficient handling of error messages in the web server. The accessDenied() method has been updated to use the error() method, and the noSuchUser() method has been removed, as it is no longer needed with the new error handling in place.